# biocontainers-api/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: GA4GH Tool Discovery API
  description: >-
    Proposed API for GA4GH (Global Alliance for Genomics & Health) tool
    repositories. A tool consists of a set of container images that are paired
    with a set of documents. Examples of documents include CWL (Common Workflow
    Language) or WDL (Workflow Description Language) or NFL (Nextflow) that
    describe how to use those images and a set of specifications for those
    images (examples are Dockerfiles or Singularity recipes) that describe how
    to reproduce those images in the future. We use the following terminology, a
    "container image" describes a container as stored at rest on a filesystem, a
    "tool" describes one of the triples as described above. In practice,
    examples of "tools" include CWL CommandLineTools, CWL Workflows, WDL
    workflows, and Nextflow workflows that reference containers in formats such
    as Docker or Singularity.
  version: 2.0.0
...
paths:
  '/tools/{id}':
    get:
      summary: 'List one specific tool, acts as an anchor for self references'
      operationId: toolsIdGet
      description: >-
        This endpoint returns one specific tool (which has ToolVersions nested
        inside it)
      tags:
        - GA4GH
      parameters:
        - name: id
          in: path
          required: true
          description: >-
            A unique identifier of the tool, scoped to this registry, for
            example `123456`
          schema:
            type: string
      responses:
        '200':
          description: A tool.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tool'
            text/plain:
              schema:
                $ref: '#/components/schemas/Tool'
        '404':
          description: The tool can not be found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            text/plain:
              schema:
                $ref: '#/components/schemas/Error'
  '/tools/{id}/versions':
    get:
      summary: List versions of a tool
      operationId: toolsIdVersionsGet
      description: Returns all versions of the specified tool
      tags:
        - GA4GH
      parameters:
        - name: id
          in: path
          required: true
          description: >-
            A unique identifier of the tool, scoped to this registry, for
            example `123456`
          schema:
            type: string
      responses:
        '200':
          description: An array of tool versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolVersion'
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolVersion'
  '/tools/{id}/versions/{version_id}':
    get:
      summary: 'List one specific tool version, acts as an anchor for self references'
      operationId: toolsIdVersionsVersionIdGet
      description: This endpoint returns one specific tool version
      tags:
        - GA4GH
      parameters:
        - name: id
          in: path
          required: true
          description: >-
            A unique identifier of the tool, scoped to this registry, for
            example `123456`
          schema:
            type: string
        - name: version_id
          in: path
          required: true
          description: >-
            An identifier of the tool version, scoped to this registry, for
            example `v1`
          schema:
            type: string
      responses:
        '200':
          description: A tool version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolVersion'
            text/plain:
              schema:
                $ref: '#/components/schemas/ToolVersion'
        '404':
          description: The tool can not be found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            text/plain:
              schema:
                $ref: '#/components/schemas/Error'
  /tools:
    get:
      summary: List all tools
      operationId: toolsGet
      description: >
        This endpoint returns all tools available or a filtered subset using
        metadata query parameters.
      tags:
        - GA4GH
      parameters:
        - name: id
          in: query
          description: >-
            A unique identifier of the tool, scoped to this registry, for
            example `123456`
          schema:
            type: string
        - name: registry
          in: query
          description: The image registry that contains the image.
          schema:
            type: string
        - name: organization
          in: query
          description: The organization in the registry that published the image.
          schema:
            type: string
        - name: name
          in: query
          description: The name of the image.
          schema:
            type: string
        - name: toolname
          in: query
          description: The name of the tool.
          schema:
            type: string
        - name: description
          in: query
          description: The description of the tool.
          schema:
            type: string
        - name: author
          in: query
          description: >-
            The author of the tool (TODO a thought occurs, are we assuming that
            the author of the CWL and the image are the same?).
          schema:
            type: string
        - $ref: '#/components/parameters/offset'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: An array of Tools that match the filter.
          headers:
            next_page:
              description: >-
                A URL that can be used to reach the next page based on the
                current offset and page record limit
              schema:
                type: string
            last_page:
              description: >-
                A URL that can be used to reach the last page based on the
                current page record limit
              schema:
                type: string
            self_link:
              description: A URL that can be used to return to the current page later
              schema:
                type: string
            current_offset:
              description: The current start index of the paging used for this result
              schema:
                type: string
            current_limit:
              description: The current page record limit used for this result
              schema:
                type: integer
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tool'
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tool'
  '/tools/{id}/versions/{version_id}/{type}/descriptor':
    get:
      summary: Get the tool descriptor for the specified tool
      operationId: toolsIdVersionsVersionIdTypeDescriptorGet
      description: >-
        Returns the descriptor for the specified tool (examples include CWL,
        WDL, or Nextflow documents).
      tags:
        - GA4GH
      parameters:
        - name: type
          required: true
          in: path
          description: >-
            The output type of the descriptor. If not specified, it is up to the
            underlying implementation to determine which output type to return.
            Plain types return the bare descriptor while the "non-plain" types
            return a descriptor wrapped with metadata. Allowable values include
            "CWL", "WDL", "NFL", "PLAIN_CWL", "PLAIN_WDL", "PLAIN_NFL".
          schema:
            type: string
        - name: id
          in: path
          description: >-
            A unique identifier of the tool, scoped to this registry, for
            example `123456`
          required: true
          schema:
            type: string
        - name: version_id
          in: path
          required: true
          description: >-
            An identifier of the tool version, scoped to this registry, for
            example `v1`
          schema:
            type: string
      responses:
        '200':
          description: The tool descriptor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolDescriptor'
            text/plain:
              schema:
                $ref: '#/components/schemas/ToolDescriptor'
        '404':
          description: The tool descriptor can not be found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            text/plain:
              schema:
                $ref: '#/components/schemas/Error'
  '/tools/{id}/versions/{version_id}/{type}/descriptor/{relative_path}':
    get:
      summary: Get additional tool descriptor files relative to the main file
      operationId: toolsIdVersionsVersionIdTypeDescriptorRelativePathGet
      description: >-
        Descriptors can often include imports that refer to additional
        descriptors. This returns additional descriptors for the specified tool
        in the same or other directories that can be reached as a relative path.
        This endpoint can be useful for workflow engine implementations like
        cwltool to programmatically download all the descriptors for a tool and
        run it
      tags:
        - GA4GH
      parameters:
        - name: type
          in: path
          required: true
          description: >-
            The output type of the descriptor. If not specified, it is up to the
            underlying implementation to determine which output type to return.
            Plain types return the bare descriptor while the "non-plain" types
            return a descriptor wrapped with metadata. Allowable values are
            "CWL", "WDL", "NFL", "PLAIN_CWL", "PLAIN_WDL", "PLAIN_NFL".
          schema:
            type: string
        - name: id
          in: path
          description: >-
            A unique identifier of the tool, scoped to this registry, for
            example `123456`
          required: true
          schema:
            type: string
        - name: version_id
          in: path
          required: true
          description: >-
            An identifier of the tool version for this particular tool registry,
            for example `v1`
          schema:
            type: string
        - name: relative_path
          in: path
          required: true
          description: >-
            A relative path to the additional file (same directory or
            subdirectories), for example 'foo.cwl' would return a 'foo.cwl' from
            the same directory as the main descriptor. 'nestedDirectory/foo.cwl'
            would return the file  from a nested subdirectory.  Unencoded paths
            such 'sampleDirectory/foo.cwl' should also be allowed
          schema:
            type: string
            pattern: .+
      responses:
        '200':
          description: The tool descriptor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolDescriptor'
            text/plain:
              schema:
                $ref: '#/components/schemas/ToolDescriptor'
        '404':
          description: The tool can not be output in the specified type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            text/plain:
              schema:
                $ref: '#/components/schemas/Error'
  '/tools/{id}/versions/{version_id}/{type}/tests':
    get:
      summary: Get a list of test JSONs
      operationId: toolsIdVersionsVersionIdTypeTestsGet
      description: >-
        Get a list of test JSONs (these allow you to execute the tool
        successfully) suitable for use with this descriptor type.
      tags:
        - GA4GH
      parameters:
        - name: type
          required: true
          in: path
          description: >-
            The type of the underlying descriptor. Allowable values include
            "CWL", "WDL", "NFL", "PLAIN_CWL", "PLAIN_WDL", "PLAIN_NFL". For
            example, "CWL" would return an list of ToolTests objects while
            "PLAIN_CWL" would return a bare JSON list with the content of the
            tests.
          schema:
            type: string
        - name: id
          in: path
          description: >-
            A unique identifier of the tool, scoped to this registry, for
            example `123456`
          required: true
          schema:
            type: string
        - name: version_id
          in: path
          required: true
          description: >-
            An identifier of the tool version for this particular tool registry,
            for example `v1`
          schema:
            type: string
      responses:
        '200':
          description: The tool test JSON response.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolTests'
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolTests'
        '404':
          description: The tool can not be output in the specified type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            text/plain:
              schema:
                $ref: '#/components/schemas/Error'
  '/tools/{id}/versions/{version_id}/{type}/files':
    get:
      summary: Get a list of objects that contain the relative path and file type
      description: >-
        Get a list of objects that contain the relative path and file type. The
        descriptors are intended for use with the
        /tools/{id}/versions/{version_id}/{type}/descriptor/{relative_path}
        endpoint.
      operationId: toolsIdVersionsVersionIdTypeFilesGet
      tags:
        - GA4GH
      parameters:
        - name: type
          required: true
          in: path
          description: >-
            The output type of the descriptor. Examples of allowable values are
            "CWL", "WDL", and "NextFlow."
          schema:
            type: string
        - name: id
          in: path
          description: >-
            A unique identifier of the tool, scoped to this registry, for
            example `123456`
          required: true
          schema:
            type: string
        - name: version_id
          in: path
          required: true
          description: >-
            An identifier of the tool version for this particular tool registry,
            for example `v1`
          schema:
            type: string
      responses:
        '200':
          description: The array of File JSON responses.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolFile'
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolFile'
        '404':
          description: The tool can not be output in the specified type.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            text/plain:
              schema:
                $ref: '#/components/schemas/Error'
  '/tools/{id}/versions/{version_id}/containerfile':
    get:
      summary: Get the container specification(s) for the specified image.
      operationId: toolsIdVersionsVersionIdContainerfileGet
      description: >-
        Returns the container specifications(s) for the specified image. For
        example, a CWL CommandlineTool can be associated with one specification
        for a container, a CWL Workflow can be associated with multiple
        specifications for containers
      tags:
        - GA4GH
      parameters:
        - name: id
          in: path
          description: >-
            A unique identifier of the tool, scoped to this registry, for
            example `123456`
          required: true
          schema:
            type: string
        - name: version_id
          in: path
          required: true
          description: >-
            An identifier of the tool version for this particular tool registry,
            for example `v1`
          schema:
            type: string
      responses:
        '200':
          description: The tool payload.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolContainerfile'
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolContainerfile'
        '404':
          description: There are no container specifications for this tool
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
            text/plain:
              schema:
                $ref: '#/components/schemas/Error'
...
servers:
  - url: /api/ga4gh/v2
components:
  parameters:
    limit:
      name: limit
      in: query
      description: Amount of records to return in a given page.
      schema:
        type: integer
        format: int32
        default: 1000
    offset:
      name: offset
      in: query
      description: >-
        Start index of paging. Pagination results can be based on numbers or
        other values chosen by the registry implementor (for example, SHA
        values). If this exceeds the current result set return an empty set.  If
        not specified in the request, this will start at the beginning of the
        results.
      schema:
        type: string
  schemas:
    ToolFile:
      type: object
      properties:
        path:
          type: string
          description: >-
            Relative path of the file.  A descriptor's path can be used with the
            GA4GH .../{type}/descriptor/{relative_path} endpoint
        file_type:
          type: string
          enum:
            - TEST_FILE
            - PRIMARY_DESCRIPTOR
            - SECONDARY_DESCRIPTOR
            - CONTAINERFILE
            - OTHER
...
    ToolDescriptor:
      type: object
      description: >-
        A tool descriptor is a metadata document that describes one or more
        tools.
      required:
        - type
      properties:
        type:
          $ref: '#/components/schemas/DescriptorType'
        descriptor:
          type: string
          description: The descriptor that represents this version of the tool.
        url:
          type: string
          description: >-
            Optional url to the underlying tool descriptor, should include
            version information, and can include a git hash
          example: >-
            https://raw.githubusercontent.com/ICGC-TCGA-PanCancer/pcawg_delly_workflow/ea2a5db69bd20a42976838790bc29294df3af02b/delly_docker/Delly.cwl
...
    Error:
      type: object
      required:
        - code
      properties:
        code:
          type: integer
          format: int32
          default: 500
        message:
          type: string
          default: Internal Server Error